--- src/App.css
@@ -1,71 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App button {
-  background-color: #007BFF;
-  font: inherit;
-  color: white;
-  border: 1px solid #007BFF;
-  padding: 8px;
-  cursor: pointer;
-}
-
-.App button:hover {
-  background-color: #0069D9;
-  color: white;
-}
-
-.App button.clicked {
-  background-color: #FFC107;
-  border: 1px solid #FFC107;
-}
-
-.App button.clicked:hover {
-  background-color: #E0A800;
-  color: black;
-}
-
-.App_logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 100px;
-}
-
-.App_header {
-  background-color: #282c34;
-  min-height: 10vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App_body {
-  margin: 5% 0 0 0;
-}
-
-.alert {
-  color: red;
-}
-
-.danger {
-  font-weight: bold;
-}
-
-
-
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
