+++ src/components/Header/Header.css
@@ -1,0 +1,32 @@
+.Header {
+    background-color: #282c34;
+    min-height: 10vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+}
+
+.Header_logo {
+    animation: App-logo-spin infinite 20s linear;
+    height: 100px;
+}
+
+.alert {
+    color: red;
+}
+  
+.danger {
+    font-weight: bold;
+}
+
+@keyframes App-logo-spin {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
