+++ src/components/Persons.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import Person from './Persons/Person/Person';
+
+const Persons = (props) => ( 
+    props.persons.map((person, index) => {
+        return <Person
+            key={person.id}
+            click={() => props.clicked(index)} 
+            name={person.name} 
+            age={person.age}
+            changed={(event) => props.changed(event, person.id)} />
+      })
+);
+
+export default Persons;
