+++ src/containers/App.css
@@ -1,0 +1,7 @@
+.App {
+  text-align: center;
+}
+
+.App-link {
+  color: #61dafb;
+}
